sqlite_to_postgres/pyproject.toml
[tool.poetry]
name = "sqlite_to_postgres"
version = "0.1.0"
description = ""
authors = ["BLARCHER"]
[tool.poetry.dependencies]
python = "^3.9"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"